# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "用户登录注册/认证示例 Golang语言编写 Gin框架",
        "title": "用户登录注册/认证示例",
        "termsOfService": "https://github.com/sun-wenming/gin-auth",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://github.com/sun-wenming/gin-auth/master/LICENSE"
        }
    },
    "host": "localhost:8000",
    "paths": {
        "/api/v1/test": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "test"
                ],
                "summary": "测试认证",
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"data\":{},\"msg\":\"ok\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "登录",
                "parameters": [
                    {
                        "description": "用户信息",
                        "name": "auth",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.auth"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"data\":{},\"msg\":\"ok\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "注册新用户",
                "parameters": [
                    {
                        "description": "用户信息",
                        "name": "auth",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.auth"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"data\":{},\"msg\":\"ok\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.auth": {
            "type": "object",
            "required": [
                "username",
                "password"
            ],
            "properties": {
                "password": {
                    "description": "PassWord 密码\n",
                    "type": "string",
                    "example": "zhangsan"
                },
                "username": {
                    "description": "UserName 用户名\n",
                    "type": "string",
                    "example": "zhangsan"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}